requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
funcsigs==1.0.2
gast==0.2.2
google-pasta==0.1.7
graphviz==0.11
grpcio==1.22.0
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.4
nltk==3.4.4
numpy==1.16.4
opencv-python==4.1.0.25
paddlepaddle-gpu==1.5.0.post97
pandas==0.24.2
Pillow==6.1.0
prettytable==0.7.2
protobuf==3.8.0
py-cpuinfo==5.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
rarfile==3.0
recordio==0.1.7
requests==2.22.0
scipy==1.2.1
six==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
funcsigs==1.0.2
gast==0.2.2
google-pasta==0.1.7
graphviz==0.11
grpcio==1.22.0
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.4
nltk==3.4.4
numpy==1.16.4
opencv-python==4.1.0.25
paddlepaddle-gpu==1.5.0.post97
pandas==0.24.2
Pillow==6.1.0
prettytable==0.7.2
protobuf==3.8.0
py-cpuinfo==5.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
rarfile==3.0
recordio==0.1.7
requests==2.22.0
scipy==1.2.1
six==1.12.0
tensorboard==1.12.2
tensorflow-estimator==1.14.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
